long to size_t

+++ philo/monitoring.c
@@ -6,13 +6,13 @@
 /*   By: molasz-a <molasz-a@student.42barcelona.co  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/02 14:34:06 by molasz-a          #+#    #+#             */
-/*   Updated: 2024/06/02 17:29:15 by molasz-a         ###   ########.fr       */
+/*   Updated: 2024/06/02 17:53:18 by molasz-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philosophers.h"
 
-static void	read_philo(t_data *data, int i, long *last_eat, int *eats)
+static void	read_philo(t_data *data, int i, size_t *last_eat, size_t *eats)
 {
 	pthread_mutex_lock(&data->philos[i].last_eats_mutex);
 	*last_eat = data->philos[i].last_eat;
@@ -31,15 +31,15 @@ static void	thread_exit(t_data *data, int i)
 	if (i >= 0)
 		print_died(&data->philos[i]);
 	else
-		printf("All philos eat at least %d times\n", data->args->min_eats);
+		printf("All philos eat at least %ld times\n", data->args->min_eats);
 	exit(0);
 }
 
 int	monitoring(t_data *data)
 {
-	long	last_eat;
-	long	time;
-	int		eats;
+	size_t	last_eat;
+	size_t	time;
+	size_t	eats;
 	int		all_eats;
 	int		i;
