+++ philo/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: molasz-a <molasz-a@student.42barcelona.co  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 20:07:30 by molasz-a          #+#    #+#             */
-/*   Updated: 2024/06/02 17:36:46 by molasz-a         ###   ########.fr       */
+/*   Updated: 2024/06/02 17:58:09 by molasz-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,11 +22,11 @@
 typedef struct s_args
 {
 	int		philos;
-	int		time_die;
-	int		time_eat;
-	int		time_sleep;
-	int		min_eats;
-	long	start;
+	size_t	time_die;
+	size_t	time_eat;
+	size_t	time_sleep;
+	size_t	min_eats;
+	size_t	start;
 }	t_args;
 
 typedef struct s_mutex
@@ -41,7 +41,7 @@ typedef struct s_philo
 	pthread_t		thread;
 	int				id;
 	int				eats;
-	long			last_eat;
+	size_t			last_eat;
 	t_args			*args;
 	t_mutex			*mutex;
 	pthread_mutex_t	eats_mutex;
@@ -62,14 +62,14 @@ void	*philo_routine(void *philo);
 int		monitoring(t_data *data);
 
 // Utils
-void	ft_sleep(long mili);
-long	get_time(void);
+void	ft_sleep(size_t mili);
+size_t	get_time(void);
 
 // Print
 void	putchar_fd(char c, int fd);
 void	print(char *str, int fd);
-void	putnbr_fd(long nb, int fd);
-void	print_time_philo(long time, int philo);
+void	putnbr_fd(size_t nb, int fd);
+void	print_time_philo(size_t time, int philo);
 void	print_fork(t_philo *philo);
 void	print_eat(t_philo *philo);
 void	print_sleep(t_philo *philo);
